confluence_requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
markdown>=3.5.0
Pygments>=2.17.0
atlassian-python-api>=3.41.0
